Integrating with supabase for data persistence

+++ apac_pipeline/storage/schema.sql
@@ -2,10 +2,10 @@ CREATE TABLE IF NOT EXISTS raw_prices (
     date        DATE NOT NULL,
     ticker      VARCHAR NOT NULL,
     market      VARCHAR NOT NULL,
-    open        DOUBLE,
-    high        DOUBLE,
-    low         DOUBLE,
-    close       DOUBLE,
+    open        DOUBLE PRECISION,
+    high        DOUBLE PRECISION,
+    low         DOUBLE PRECISION,
+    close       DOUBLE PRECISION,
     volume      BIGINT,
     PRIMARY KEY (date, ticker)
 );
@@ -14,12 +14,12 @@ CREATE TABLE IF NOT EXISTS adjusted_prices (
     date        DATE NOT NULL,
     ticker      VARCHAR NOT NULL,
     market      VARCHAR NOT NULL,
-    open        DOUBLE,
-    high        DOUBLE,
-    low         DOUBLE,
-    close       DOUBLE,
+    open        DOUBLE PRECISION,
+    high        DOUBLE PRECISION,
+    low         DOUBLE PRECISION,
+    close       DOUBLE PRECISION,
     volume      BIGINT,
-    adj_factor  DOUBLE,
+    adj_factor  DOUBLE PRECISION,
     PRIMARY KEY (date, ticker)
 );
 
@@ -27,7 +27,7 @@ CREATE TABLE IF NOT EXISTS corporate_actions (
     date        DATE NOT NULL,
     ticker      VARCHAR NOT NULL,
     action_type VARCHAR NOT NULL,
-    value       DOUBLE NOT NULL,
+    value       DOUBLE PRECISION NOT NULL,
     PRIMARY KEY (date, ticker, action_type)
 );
 
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS macro_indicators (
     market      VARCHAR NOT NULL,
     indicator   VARCHAR NOT NULL,
     source      VARCHAR NOT NULL,
-    value       DOUBLE NOT NULL,
+    value       DOUBLE PRECISION NOT NULL,
     PRIMARY KEY (date, market, indicator)
 );
